Provide initial functionality for creating links in an existing graph
Provide a new mode (LINK) for the cli interface which can be used to create links on existing graphs using a txt file exported through depthmapX (by clicking the link button and then Map -> Export). Uses existing depthmapX functionality as much as possible i.e. passes the text stream to a temporary ShapeMap which will in trun generate a list of Lines. The coordinates of the ends are then passed to pixelate() to get the relevant grid pixels which are paired and used to trigger mergePixels(). This commit uses a placeholder string which works for the testdata/gallery_connected.graph file instead of an exported txt file