When zooming get the centre of the window as in 'LogicalUnits' to avoid drift
The function this->rect().center() gives an x value of 531 instead of 533 for a window of width 1066 pixels. This only happens for even window widths. In QDepthmapView::LogicalUnits the centre is instead calculated by dividing the width by 2 which gives the correct response and thus (currently) it causes the whole plan to drift. Since the wheelEvent function calls LogicalUnits it should at least calculate the centre the same way