Change ViewHelpers::calculateCenter to output Point2f
When calculating the new centre for zooming in/out it makes more sense to leave the calculations in continuous space (QPoint only takaes integers x and y). If not, the rounding that happens forces the plan to drift along the screen as the user zooms in and out. The expected behaviour would be for the mouse to stay at the same spot on the plan before and after the zooming, despite the scale of the plan changing. QDepthmapView::LogicalUnits can not be used in the process as it only accepts QPoints and is thus purely for translating direct pixel input (i.e. a user clicks on a spot on the screen). Its functionality has been replicated in the QDepthmapView::wheelEvent function