Make sure changes are propagated to the GLView
The function paintGL for GLView is not on a timer and will only update whenever something changes. Therefore there is no way to constantly check the redraw flags in GraphDoc. Instead, a temporary hacky solution is applied here: When GraphDoc::SetRedrawFlag() is called from various parts of the program itself calls all the GLView children to be updated, thus triggering an update event and subsequently a paintGL event. Also added here a modification in GLView to catch cases where the metagraph region is not truly set (all values at 0)