Cleanup warnings (#294)
* Silence OpenGL macOS warnings

* Remove version variable from all read and write functions

* Add default case handling to all switch statements

* Remove 'const' type qualifiers that have no effect

* Remove communicator variable name where not required

* Remove unused functions

* Remove all unused parameters

* Remove unused variables

* Properly order initialisation lists

* Fix multi-line comment

* Clear ambiguity of logical operations

* Add missing virtual destructors

* Replace old finite() with isfinite()

* Remove the register specifier

* Change __max to std::max

* Use std::move for SalaErrors in SalaProgram

* Fix operator order

* Fix other minor issues

* Squash all warnings when comparing integers of different signs and sizes

* Read count as an integer

* Remove unused variable

* Remove last version signifiers

* Reimplement various glu functions

Apple's deprecation of OpenGL creates a different problem here because the 3DView is coded using the old OpenGL way. Once 3DView is adapted to work as the GLView (or merged into it) then the glureimpl.h file won't be required. The license of the original file and thus the code in glureimpl.h is the SGI Free Sofware License B, version 2 and it is compatible with the GPL according to https://www.gnu.org/licenses/license-list.html

* Remove unnecessary code

* Remove unused functions and reduntant file

* Add missing math declarations

* Remove non-existant file reference

* Squash some gcc warnings

* Restore and expose a hidden colourscale

The colourscale g_nicecolorhsb along with the relevant comment "Test a range designed to try to keep consistent saturation and brightness of g_nicecolor, and only move hue" has been named "Hue Only Axmanesque" as it's a variation of that colourscale. On the interface it appears as "Equal Ranges (3-Colour Hue Only)"

* Remove qt5_wrap_cpp as AUTOMOC is set to ON